Define enum for pre and post orders
Recursive data structures can be iterated in pre or post order.
In many places, the order was kept as a string with either 'pre'
or 'post'. A better solution is to use an enum to keep the order.

This commit adds an enumeration named Order with two values PRE
and POST.