Add support for pruning
While iterating over recursive data structures, it may be
desirable to prune certain instances and their subinstances.
The current implementation of iterators does not support this
feature and requires filtering to achieve it.

This commit adds support for pruning functions for instances of
both Recursive and MultiRecursive. An example was also added to
illustrate this new feature.