Streamline the recur.tree.Tree class
The example Tree class provided by the package made assumptions
about its data instead of simply managing the ADT. It was also
used in the tests of the abc module which caused problems
for maintenance.

This commit removes the value attribute of the Tree class and
updates the tests and examples accordingly. It also removes the
tests using the Tree class in the test_abc module.