Add direction of traversal
The direction of traversal is selected for the user and is left to
right in preorder and right to left in postorder. Most of the
existing literature indicates that postorder should also traverse
the structure in the left to right direction.

This commit adds the notion of direction to the traversal of
structures. By default, it is now always from left to right, but
can be modified on demand.